requirements.txt
arrow==0.17.0
cartopy==0.18.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastapi==0.61.1
h11==0.11.0
idna==2.10
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
polyline==1.4.0
pydantic==1.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
starlette==0.13.6
stravalib==0.10.2
typing-extensions==3.7.4.3
units==0.7
urllib3==1.25.11
uvicorn==0.12.2
